Fix issue with added partitions
This could happen:
- the processor starts with 2 partitions p0 and p1
- p0 has some item, p1 doesn't
- the processor saves a snapshot. It will only save the offset for p0
because it doesn't know it for p1
- the job restarts
- the processor restores offset for p0, but not for p1. Therefore it
will assume the partition count before the restart was 1
- later it will detect that there are 2 partitions and will seek p1 to
the beginning because it will think that p1 was added while the job was
down

The fix is that the processor with global processor index of 0 will save
one additional key to the snapshot, the `PARTITION_COUNTS_SNAPSHOT_KEY`.
Under that key we the known partition count for each topic. That key is
broadcast, every processor will know the partition count before the
restart. After this, the above scenario will not consider p1 as added
and will not seek to the beginning.