Fix when_partitionAddedWhileJobDown test failure
This happened:

- assertTrueEventually calls the given lambda repeatedly, until no
assert fails. Therefore it could produce multiple records. This is
intentional because our auto.offset.reset is latest and we don't know
where the consumer starts

- on one iteration a record is produced, but sinkList is still empty (it
didn't yet made its way through the pipeline). On a next iteration,
another record was produced, but sinkList contained 1 item, so the
`assertTrueEventually` method returned

- then we suspended the job and cleared the list. However, `job.suspend`
returns before the execution actually terminates. We cleared the list,
but the job was still running and inserted one more item into it

- the test then failed because of an unexpected item in the list

The fix is to not rely on a cleared list, but just check the last item
in the `sinkList`.

Fixes #2747