Fix race in snapshot phase 1 completion
There was no locking in
`MasterSnapshotContext.onSnapshotPhase1Complete`. This was possible:

- snapshot 1 is the last successful one

- snapshot 2 was initiated, the phase 1 was successful on member.
`onSnapshotPhase1Complete` was entered on thread `A`. Current
executionId and the executionId from the response were equal

- in another threads, the execution was forcibly restarted,
`finalizeJob` called, new execution started and started and was about to
restore from snapshot 1 (the snapshot 2 wasn't yet marked as successful
by `onSnapshotPhase1Complete`)

- thread `A` continued assuming that snapshot 2 is the new successful
one, cleared the map for snapshot 1

- snapshot 1 was restored from an cleared map - state wasn't restored.
The source started emission from 0 and caused the duplicate results.

Many other scenarios are possible, this is just what happened in the
test failure I'm claiming to fix by this PR. Possibly (hopefully?) some
of the _stuck snapshot_ issues were also caused by this.

Fixes #2504